Moved App component
