+++ src/components/App/App.tsx
@@ -1,14 +1,12 @@
 import React from "react";
 import "./App.css";
-import Chat from "./components/Chat/Chat";
+import Chat from "../Chat/Chat";
 
-function App() {
+export const App = () => {
   return (
     <div className="app">
       <h1 className="title">Simple Chat Maquette</h1>
       <Chat />
     </div>
   );
-}
-
-export default App;
+};
